requirements.txt
asgiref==3.6.0
cachetools==5.3.0
charset-normalizer==3.1.0
Django==4.1.7
django-environ==0.10.0
django-storages==1.13.2
djangorestframework==3.14.0
drf-spectacular==0.26.1
google-api-core==2.11.0
google-auth==2.17.0
google-cloud-core==2.3.2
google-cloud-secret-manager==2.16.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.53.0
grpcio-status==1.53.0
gunicorn==20.1.0
idna==3.4
proto-plus==1.22.2
protobuf==4.22.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2023.2
requests==2.28.2
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tzdata==2023.2
urllib3==1.26.15
wincertstore==0.2
